Tried to use Stimulus
I could not get Stimulus controllers  working so I tried to revert to my original JS code, but that stopped working, so I just put script tags in the  appropriate erb files and carried on. Added users to reviews and added extra fields to the User model to make the displayed reviews more interesting.